requirements.txt
configparser==3.7.1
falcon==1.4.1
falcon-jsonify==1.2
falcon-autocrud==1.0.26
gunicorn==19.9.0
marshmallow==2.18.0
psycopg2-binary==2.7.7
python-mimeparse==1.6.0
six==1.12.0
SQLAlchemy==1.2.17
ujson==1.35
pyjwt==1.7.1
bcrypt==3.1.6
